moved run logic to cli::run
